+++ bin/loader/lib/plugin-loader.js
