+++ bin/loader/plugins/prune.js
