+++ bin/loader/plugins/stats.js
