+++ bin/methods/start.js
@@ -55,7 +55,7 @@ module.exports = class Cordlr {
     this.bot.bin = this
 
     // Initiate the Loader
-    const Loader = require(path.join(__dirname, '../../loader/index'))
+    const Loader = require(path.join(__dirname, '../loader/index'))
     return new Loader(this.bot, this.config)
   }
